+++ src/client/config_files.c
@@ -12,17 +12,45 @@
 #include "dmi.h"
 #include "str_functions.h"
 
-// get an array of config names
-array_of(ConfigFile) get_configs() {
+// Check if `files` contains a config named `name`
+bool Contains_Config(array_of(ConfigFile)* files, const char* name) {
+  for_each_array(ConfigFile*, file, *files) {
+    if (! strcmp(file->config_name, name))
+      return true;
+  }
+
+  return false;
+}
+
+// Free an array of ConfigFile
+void ConfigFiles_Free(array_of(ConfigFile)* files) {
+  for_each_array(ConfigFile*, file, *files)
+    Mem_Free(file->config_name);
+  Mem_Free(files->data);
+}
+
+// Compare function for qsort
+int compare_config_by_name(const void *a, const void *b) {
+  return strcmp(((struct ConfigFile *)a)->config_name, ((struct ConfigFile *)b)->config_name);
+}
+
+// Compare function for qsort
+int compare_config_by_diff(const void *a, const void *b) {
+  return (((struct ConfigFile *)b)->diff > ((struct ConfigFile *)a)->diff)
+       - (((struct ConfigFile *)b)->diff < ((struct ConfigFile *)a)->diff);
+}
+
+// Return an array of ConfigFile for each file in `path`
+static array_of(ConfigFile) List_Configs_In_Directory(const char* path) {
   size_t capacity = 512;
   array_of(ConfigFile) files = {
-    .data = Mem_Malloc(capacity * sizeof(struct ConfigFile)),
+    .data = Mem_Malloc(capacity * sizeof(ConfigFile)),
     .size = 0
   };
 
-  DIR* directory = opendir(NBFC_MODEL_CONFIGS_DIR);
+  DIR* directory = opendir(path);
   if (!directory) {
-    Log_Error("Failed to open directory `" NBFC_MODEL_CONFIGS_DIR "': %s\n", strerror(errno));
+    Log_Error("Failed to open directory `%s': %s\n", path, strerror(errno));
     exit(NBFC_EXIT_FAILURE);
   }
 
@@ -38,7 +66,7 @@ array_of(ConfigFile) get_configs() {
 
     if (files.size == capacity) {
       capacity *= 2;
-      files.data = Mem_Realloc(files.data, capacity * sizeof(struct ConfigFile));
+      files.data = Mem_Realloc(files.data, capacity * sizeof(ConfigFile));
     }
 
     files.data[files.size++].config_name = Mem_Strdup(file->d_name);
@@ -48,95 +76,116 @@ array_of(ConfigFile) get_configs() {
   return files;
 }
 
-// compare function for qsort
-int compare_config_by_name(const void *a, const void *b) {
-  return strcmp(((struct ConfigFile *)a)->config_name, ((struct ConfigFile *)b)->config_name);
+// Merges two arrays of ConfigFile into a single array,
+// removing duplicates based on `config_name`.
+static array_of(ConfigFile) Merge_Configs(array_of(ConfigFile)* a, array_of(ConfigFile)* b) {
+  array_of(ConfigFile) files = {
+    .data = Mem_Malloc((a->size + b->size) * sizeof(ConfigFile)),
+    .size = 0
+  };
+
+  for_each_array(ConfigFile*, file, *a) {
+    files.data[files.size++].config_name = Mem_Strdup(file->config_name);
+  }
+
+  for_each_array(ConfigFile*, file, *b) {
+    if (! Contains_Config(&files, file->config_name)) {
+      files.data[files.size++].config_name = Mem_Strdup(file->config_name);
+    }
+  }
+
+  return files;
 }
 
-// compare function for qsort
-int compare_config_by_diff(const void *a, const void *b) {
-  return (((struct ConfigFile *)b)->diff > ((struct ConfigFile *)a)->diff)
-       - (((struct ConfigFile *)b)->diff < ((struct ConfigFile *)a)->diff);
+// List all configs (in the static config directory as well as in the mutable config directory).
+array_of(ConfigFile) List_All_Configs() {
+  array_of(ConfigFile) a = List_Configs_In_Directory(NBFC_MODEL_CONFIGS_DIR);
+  array_of(ConfigFile) b = List_Configs_In_Directory(NBFC_MODEL_CONFIGS_DIR_MUTABLE);
+  array_of(ConfigFile) c = Merge_Configs(&a, &b);
+  ConfigFiles_Free(&a);
+  ConfigFiles_Free(&b);
+  return c;
 }
 
-// recommend configs in a sorted array
-array_of(ConfigFile) recommended_configs() {
-  const char *product = get_model_name();
-  array_of(ConfigFile) files = get_configs();
+// List all configs (in the static config directory as well as in the mutable config directory).
+// The `diff` field of the ConfigFile structure will also be set.
+array_of(ConfigFile) List_Recommended_Configs() {
+  const char* model_name = get_model_name();
+  array_of(ConfigFile) files = List_All_Configs();
   for_each_array(ConfigFile*, file, files) {
-    file->diff = str_similarity(product, file->config_name);
+    file->diff = str_similarity(model_name, file->config_name);
   }
   qsort(files.data, files.size, sizeof(struct ConfigFile), compare_config_by_diff);
   return files;
 }
 
 /*
- * Retrive the supported model for `model_name`.
+ * Retrive the supported config for `model_name`.
  *
- * This function searches the model support database (a JSON file) for the
- * given `model_name` and returns the compatible output model for it.
+ * This function searches a model support database (a JSON file) for the
+ * given `model_name` and returns the compatible output config for it.
  *
  * The model support database is a JSON object where each key represents an
  * input model name and each value represents an output model name:
  *
  *  {
- *     "Input Model Name": "Output Model Name"
+ *     "Input Model Name": "Output Config"
  *  }
  *
  * If the model is found in the support database, the function will check if
- * the corresponding output model exists in the provided `config_files`.
- * If a match is found, the output model is returned. Otherwise, a warning
+ * the corresponding output config exists in the provided `config_files`.
+ * If a match is found, the output config is returned. Otherwise, a warning
  * is printed and `NULL` is returned.
  *
  * If the input model is not found in the support database, the function
- * attempts to find the model name directly in `config_files` and returns it if
+ * attempts to find the config directly in `config_files` and returns it if
  * a match is found.
  *
  * If there is no match, `NULL` is returned.
  */
-char* get_supported_model(array_of(ConfigFile)* config_files, const char* model_name) {
+char* Get_Supported_Config_From_SupportFile(const char* support_file, array_of(ConfigFile)* config_files, const char* model_name) {
   char buf[NBFC_MAX_FILE_SIZE];
   const nx_json* root = NULL;
-  char* output_model = NULL;
+  char* config = NULL;
 
-  Error* e = nx_json_parse_file(&root, buf, sizeof(buf), NBFC_MODEL_SUPPORT_FILE);
+  Error* e = nx_json_parse_file(&root, buf, sizeof(buf), support_file);
   if (e) {
-    Log_Warn("%s: %s\n", NBFC_MODEL_SUPPORT_FILE, err_print_all(e));
+    Log_Warn("%s: %s\n", support_file, err_print_all(e));
     goto end;
   }
 
   if (root->type != NX_JSON_OBJECT) {
-    Log_Warn("%s: Not a JSON object\n", NBFC_MODEL_SUPPORT_FILE);
+    Log_Warn("%s: Not a JSON object\n", support_file);
     goto end;
   }
 
   nx_json_for_each(model, root) {
     if (model->type != NX_JSON_STRING) {
-      Log_Warn("%s: Invalid value for model `%s`: Not a string\n", NBFC_MODEL_SUPPORT_FILE, model->key);
+      Log_Warn("%s: Invalid value for model `%s`: Not a string\n", support_file, model->key);
     }
     else if (!strcmp(model->key, model_name)) {
-      if (output_model) {
-        Log_Warn("%s: Duplicate model key: `%s`\n", NBFC_MODEL_SUPPORT_FILE, model->key);
+      if (config) {
+        Log_Warn("%s: Duplicate model key: `%s`\n", support_file, model->key);
       }
-      output_model = Mem_Strdup(model->val.text);
+      config = Mem_Strdup(model->val.text);
     }
   }
 
 end:
   nx_json_free(root);
 
-  if (output_model) {
+  if (config) {
     // Ensure that the model actually exists
     for_each_array(ConfigFile*, file, *config_files) {
-      if (!strcmp(file->config_name, output_model)) {
-        return output_model;
+      if (!strcmp(file->config_name, config)) {
+        return config;
       }
     }
 
     Log_Warn("%s: The model `%s` was found in the support database, but the specified configuration file (`%s`) is missing\n",
-        NBFC_MODEL_SUPPORT_FILE, model_name, output_model);
+        support_file, model_name, config);
 
-    Mem_Free(output_model);
+    Mem_Free(config);
   }
   else {
     // Not found in support database, try a direct match on `config_files`
@@ -150,3 +199,14 @@ end:
   return NULL;
 }
 
+char* Get_Supported_Config(array_of(ConfigFile)* files, const char* model) {
+  char* config = NULL;
+
+  if (access(NBFC_MODEL_SUPPORT_FILE_MUTABLE, F_OK) == 0)
+    config = Get_Supported_Config_From_SupportFile(NBFC_MODEL_SUPPORT_FILE_MUTABLE, files, model);
+
+  if (! config)
+    config = Get_Supported_Config_From_SupportFile(NBFC_MODEL_SUPPORT_FILE, files, model);
+
+  return config;
+}