+++ src/client/mkdir_p.c
@@ -1,0 +1,34 @@
+#include <sys/stat.h>  // mkdir
+#include <sys/types.h> // mode_t
+#include <errno.h>     // errno, EEXIST
+
+#include "../memory.h"
+
+int mkdir_p(const char *path, mode_t mode) {
+  char *tmp = Mem_Strdup(path);
+  char *p = tmp;
+  int ret = 0;
+
+  if (tmp[0] == '/')
+    p++;
+
+  while (*p) {
+    if (*p == '/') {
+      *p = '\0';
+      if (mkdir(tmp, mode) != 0) {
+        if (errno != EEXIST) {
+          ret = -1;
+          break;
+        }
+      }
+      *p = '/';
+    }
+    p++;
+  }
+
+  if (ret == 0 && mkdir(tmp, mode) != 0 && errno != EEXIST)
+    ret = -1;
+
+  Mem_Free(tmp);
+  return ret;
+}